refactor: move exception to entities
