+++ src/domain/entities/exception.entity.ts
